Move query and mutation implementation to relevant submodules
