+++ server/src/graphql/nodes/mod.rs
@@ -1,11 +1,7 @@
+mod node;
 mod person;
 mod squad;
 
+pub use node::*;
 pub use person::*;
 pub use squad::*;
-
-#[async_graphql::Interface(field(name = "id", type = "String"))]
-pub enum Node {
-    Person(Person),
-    Squad(Squad),
-}
